Corrected code

+++ src/aunmag/nightingale/utilities/Timer.java
@@ -40,8 +40,8 @@ public class Timer {
         );
     }
 
-    public float calculateIsDoneRatio() {
-        return (float) (getPassed() / duration);
+    public double calculateIsDoneRatio() {
+        return getPassed() / duration;
     }
 
     /* Setters */
